refactor: rename icons folder
